scripts: move more common funcs into utils
